refactor: move directory outwards
