+++ package.json
@@ -1,0 +1,71 @@
+{
+  "name": "salvaguarda",
+  "version": "1.0.0",
+  "main": "node_modules/expo/AppEntry.js",
+  "scripts": {
+    "start": "expo start",
+    "start-tunnel": "expo start --tunnel",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "test": "jest"
+  },
+  "dependencies": {
+    "@expo/metro-config": "^0.4.0",
+    "@expo/ngrok": "^4.1.0",
+    "@expo/vector-icons": "^13.0.0",
+    "@react-navigation/drawer": "^6.5.0",
+    "@react-navigation/native": "^6.0.13",
+    "@react-navigation/native-stack": "^6.9.0",
+    "expo": "~46.0.13",
+    "expo-image-picker": "~13.3.1",
+    "expo-status-bar": "~1.4.0",
+    "firebase": "^9.10.0",
+    "jest": "^28.0.0",
+    "jest-expo": "^46.0.1",
+    "react": "18.0.0",
+    "react-dom": "18.0.0",
+    "react-hook-form": "^7.36.1",
+    "react-native": "0.69.6",
+    "react-native-dropdown-picker": "^5.4.2",
+    "react-native-elements": "latest",
+    "react-native-gesture-handler": "~2.5.0",
+    "react-native-gifted-chat": "0.16.3",
+    "react-native-reanimated": "~2.9.1",
+    "react-native-safe-area-context": "3.1.9",
+    "react-native-screens": "~3.15.0",
+    "react-native-vector-icons": "^9.2.0",
+    "react-native-web": "~0.18.7"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.9",
+    "@testing-library/jest-native": "^5.1.2",
+    "@testing-library/react-native": "^11.3.0",
+    "@types/jest": "^29.2.0",
+    "@types/react": "~18.0.14",
+    "@types/react-native": "~0.69.1",
+    "@types/react-test-renderer": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "eslint": "^8.26.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-standard-with-typescript": "^23.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.3.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "react-test-renderer": "18.0.0",
+    "typescript": "^4.8.4"
+  },
+  "jest": {
+    "preset": "jest-expo",
+    "transformIgnorePatterns": [
+      "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)"
+    ],
+    "moduleFileExtensions": [
+      "ts",
+      "tsx",
+      "js",
+      "jsx"
+    ]
+  },
+  "private": true
+}
