--- salvaguarda/src/definitions/ISubscription.ts
@@ -1,6 +1,0 @@
-export type ISubscription = {
-	_id: string;
-	uid: string;
-	rid: string;
-    createdAt: Date;
-};
