--- salvaguarda/src/definitions/index.ts
@@ -1,6 +1,0 @@
-export * from './IUser';
-export * from './IRoom';
-export * from './ISubscription';
-export * from './IMessage';
-export * from './ScreenPropsTypes';
-export * from './IQuestion';
