+++ src/definitions/IMessage.ts
