+++ src/definitions/ISubscription.ts
@@ -1,0 +1,6 @@
+export type ISubscription = {
+	_id: string;
+	uid: string;
+	rid: string;
+    createdAt: Date;
+};
