+++ src/views/addImage.tsx
